+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog for AWS SDK for JavaScript
-<!--LATEST=2.517.0-->
+<!--LATEST=2.518.0-->
 <!--ENTRYINSERT-->
+
+## 2.518.0
+* bugfix: Credentials: Update environment variable names in TokenFileWebIdentityCredentials
+* feature: SSM: This feature adds "default tier" to the AWS Systems Manager Parameter Store for parameter creation and update. AWS customers can now set the "default tier" to one of the following values: Standard (default), Advanced or Intelligent-Tiering.  This allows customers to create advanced parameters or parameters in corresponding tiers with one setting rather than code change to specify parameter tiers.
+* feature: SecurityHub: This release resolves an issue with the DescribeHub action, changes the MasterId and InvitationId parameters for AcceptInvitation to Required, and changes the AccountIds parameter for DeleteInvitations and DeclineInvitations to Required.
 
 ## 2.517.0
 * feature: Credentials: Added TokenFileWebIdentityCredentials